Update PCSleepTimer.ino
added sleep flag to prevent re-triggers

+++ PCSleepTimer/PCSleepTimer.ino
@@ -2,6 +2,7 @@
 byte GNDPin = 3;
 byte VCCPin = 2;
 byte dataPin = 4;
+byte sleepSetFlag = 0;
 
 void setup() {
   pinMode(GNDPin, OUTPUT);
@@ -17,17 +18,19 @@ void setup() {
 }
 
 void loop() {
-  if(digitalRead(dataPin) == 0){
-    Keyboard.press(KEY_LEFT_GUI);
-    Keyboard.press('r');
-    Keyboard.releaseAll();
-    delay(1000);
-    Keyboard.print("shutdown -s -t 1800");
-    delay(1000);
-    Keyboard.press(KEY_RETURN);
-    delay(100);
-    Keyboard.releaseAll();
-    delay(100);
+  if(!sleepSetFlag){
+    if(digitalRead(dataPin) == 0){
+      Keyboard.press(KEY_LEFT_GUI);
+      Keyboard.press('r');
+      Keyboard.releaseAll();
+      delay(1000);
+      Keyboard.print("shutdown -s -t 1800");
+      delay(1000);
+      Keyboard.press(KEY_RETURN);
+      delay(100);
+      Keyboard.releaseAll();
+      delay(100);
+      sleepSetFlag = 1;
+    }
   }
-  
 }
